+++ src/optimum/rbln/ops/flash_attn.py
@@ -44,16 +44,16 @@ def register_rbln_custom_flash_attention():
     def flash_attn_decode_cpu(q, k, v, mask, kcache, vcache, seq, scale, partition):
         return (
             q,
-            torch.empty(1, *kcache.shape[1:], device=kcache.device),
-            torch.empty(1, *vcache.shape[1:], device=vcache.device),
+            torch.empty(1, *kcache.shape[1:], device=kcache.device, dtype=kcache.dtype),
+            torch.empty(1, *vcache.shape[1:], device=vcache.device, dtype=vcache.dtype),
         )
 
     @register_fake("rbln_custom_ops::flash_attn_decode")
     def flash_attn_decode_abstract(q, k, v, m, kcache, vcache, seq, scale, partition):
         return (
             q,
-            torch.empty(1, *kcache.shape[1:], device=kcache.device),
-            torch.empty(1, *vcache.shape[1:], device=vcache.device),
+            torch.empty(1, *kcache.shape[1:], device=kcache.device, dtype=kcache.dtype),
+            torch.empty(1, *vcache.shape[1:], device=vcache.device, dtype=vcache.dtype),
         )
 
     torch.library.define(
